Revert "fixed"
This reverts commit da86935f1dd8c3deffcffff9ec5ece6b7abd55ca.

+++ repository/src/main/java/com/epam/esm/mapper/TagMapper.java
@@ -18,7 +18,7 @@ public interface TagMapper {
     long save(Tag tag);
 
     @Delete("DELETE FROM tags WHERE id = #{id}")
-    int delete(long id);
+    int delete(int id);
 
     @Select("SELECT EXISTS(SELECT name FROM tags WHERE name = #{name})")
     Boolean existByName(String name);
